Simple bitwise allocator

+++ kernel/utils/bitops.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // bit set from the right, bval either 0 or 1
-#define BIT_SET(val, bitnum, bval) ((val & ~(1 << bitnum)) | (bval << bitnum))
+#define BIT_SET(val, bitnum, bval) ((val & ~(1ull << bitnum)) | (((uint64)bval) << bitnum))
